@@ -10,18 +10,18 @@ description: Demonstrates how to use the image_gallery_saver plugin.
 version: 1.5.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.13+hotfix.5 <2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  permission_handler: ^5.0.1
-  fluttertoast: ^4.0.0
+  permission_handler: ^6.0.1
+  fluttertoast: ^8.0.1-nullsafety.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
@@ -30,8 +30,8 @@ dev_dependencies:
   image_gallery_saver:
     path: ../
 
-  path_provider: ^1.6.11
-  dio: ^3.0.9
+  path_provider: ^2.0.1
+  dio: ^4.0.0-beta6
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
