@@ -5,7 +5,7 @@ author: quicey <quiceyxu@gmail.com>
 homepage: https://github.com/hui-z/image_gallery_saver
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.13+hotfix.5 <2.0.0"
 
 dependencies:
